edit main page

+++ src/main/java/com/leka/blogteashop/service/impl/PostServiceImpl.java
@@ -11,13 +11,15 @@ import com.leka.blogteashop.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.core.io.ByteArrayResource;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.client.MultipartBodyBuilder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.Iterator;
 import java.util.List;
-import java.util.regex.Pattern;
 
 @Service
 @RequiredArgsConstructor
@@ -33,11 +35,8 @@ public class PostServiceImpl implements PostService {
     @Transactional
     public PostResponseOnlyId addPost(PostDto postDto, MultipartFile bgFile, List<MultipartFile> postFiles) {
         Post post = postMapper.toEntity(postDto);
-        MultipartBodyBuilder builder;
         if (bgFile != null && !bgFile.isEmpty()) {
-            builder = createFrom(bgFile, true);
-            ImageDto bgImage = mediaService.uploadImageThrough(builder);
-            post.addImage(new Image(bgImage.getId(), bgImage.getFileName(), post));
+            uploadBackgroundImageToPost(bgFile, post);
         }
         addPostImages(postFiles, post);
         return postMapper.toResponseOnlyId(postRepository.save(post));
@@ -59,37 +58,52 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void editPost(Long postId, EditPostDto postDto, MultipartFile bgImage, List<MultipartFile> postFiles) {
+    public void editPost(Long postId, EditPostDto postDto, MultipartFile bgFile, List<MultipartFile> postFiles) {
         Post post = getPost(postId);
         postMapper.updatePost(post, postDto);
-        List<Image> imagesBeforeUpdate = post.getPostImages();
+        List<Image> postImages = post.getPostImages();
         MultipartBodyBuilder builder;
-        if (bgImage != null && !bgImage.isEmpty()) {
-            for (Image image: imagesBeforeUpdate) {
+        if (bgFile != null && !bgFile.isEmpty()) {
+            boolean isUpdated = false;
+            for (Image image: postImages) {
                 if (image.getImageName().startsWith(PREFIX_BACKGROUND_IMAGE)) {
-                    builder = createFrom(bgImage, true);
+                    builder = createFrom(bgFile, true);
                     Long imageId = image.getImageId();
-                    mediaService.updateImageById(imageId, builder);
+                    ImageDto imageDto = mediaService.updateImageById(imageId, builder);
+                    image.setImageName(imageDto.getFileName());
+                    isUpdated = true;
                     break;
                 }
             }
+            if (!isUpdated) {
+                uploadBackgroundImageToPost(bgFile, post);
+            }
         }
         //delete old files by checking that the image name is not found in updated contents of EditPostDto
-        for (Image image: imagesBeforeUpdate) {
+        Iterator<Image> iterator = postImages.iterator();
+        while (iterator.hasNext()) {
+            Image image = iterator.next();
+            if (image.getImageName().startsWith(PREFIX_BACKGROUND_IMAGE)) continue;
             String target = "@{%s}".formatted(image.getImageName());
             //in order to delete old images from database, we have to check that all contents in different languages
-            //don't include references to image names, if the admin forgets to delete in English version for example,
+            //don't include references to image names. If the admin forgets to delete in English version for example,
             // the image won't be deleted in general and will be shown in English content only.
             if (!StringUtils.contains(postDto.getContentUA(), target)
                     && !StringUtils.contains(postDto.getContentEN(), target) ){
                 Long imageId = image.getImageId();
                 mediaService.deleteImageById(imageId);
-                post.removeImage(image);
+                iterator.remove();
             }
         }
         //upload new ones if present
         addPostImages(postFiles, post);
+        post.setPostImages(postImages);
         postRepository.save(post);
+    }
+
+    @Override
+    public Page<PostResponse> getPosts(Pageable pageable) {
+        return postRepository.findAll(pageable).map(postMapper::toResponse);
     }
 
     private MultipartBodyBuilder createFrom(MultipartFile file, boolean isBackground) {
@@ -124,6 +138,12 @@ public class PostServiceImpl implements PostService {
     private Post getPost(Long postId) {
         return postRepository.findById(postId)
                 .orElseThrow(() -> new NotFoundException("Post not found"));
+    }
+
+    private void uploadBackgroundImageToPost(MultipartFile bgFile, Post post) {
+        MultipartBodyBuilder builder = createFrom(bgFile, true);
+        ImageDto bgImage = mediaService.uploadImageThrough(builder);
+        post.addImage(new Image(bgImage.getId(), bgImage.getFileName(), post));
     }
 
 }
